core/requirements.txt
gunicorn==20.1.0
PyYAML==6.0
pony==0.7.16